Refactor: shared 2D interleaved/sharded matmul helper + parameterized bias_dtype
- Move _run_matmul_2d_interleaved_in0_sharded_in1 to test_matmul_deepseek.py;
  use it in test_prefill_mm_interleaved_sharded and in nightly test_matmul.py.
- Add bias_dtype param to helper; parameterize bias_dtype in dtype sweep test.
